+++ src/main/java/br/ufsc/bridge/res/dab/dto/ResABIdentificacaoProfissional.java
@@ -2,6 +2,7 @@ package br.ufsc.bridge.res.dab.dto;
 
 import javax.xml.xpath.XPathExpressionException;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,6 +12,7 @@ import br.ufsc.bridge.res.util.XPathFactoryAssist;
 @Getter
 @Setter
 @NoArgsConstructor
+@EqualsAndHashCode
 public class ResABIdentificacaoProfissional {
 
 	private String cns;
