+++ src/main/java/br/ufsc/bridge/res/http/ResHttpClient.java
@@ -15,6 +15,7 @@ import javax.xml.soap.SOAPException;
 
 import lombok.extern.slf4j.Slf4j;
 
+import org.apache.commons.io.IOUtils;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
@@ -44,21 +45,13 @@ public class ResHttpClient {
 	public ResHttpClient(CreateSOAPMessage soapMessage, String action) {
 		this.soapMessage = soapMessage;
 
-		RequestConfig requestConfig = RequestConfig.custom()
-				.setSocketTimeout(60000)
-				.setConnectTimeout(30000)
-				.setExpectContinueEnabled(false)
-				.setRedirectsEnabled(true)
-				.build();
+		RequestConfig requestConfig = RequestConfig.custom().setSocketTimeout(60000).setConnectTimeout(30000).setExpectContinueEnabled(false).setRedirectsEnabled(true).build();
 
 		PoolingHttpClientConnectionManager cm = new PoolingHttpClientConnectionManager();
 		cm.setValidateAfterInactivity(10000);
 		cm.setDefaultSocketConfig(SocketConfig.custom().setTcpNoDelay(true).build());
 
-		this.httpClient = HttpClients.custom()
-				.setDefaultRequestConfig(requestConfig)
-				.setConnectionManager(cm)
-				.build();
+		this.httpClient = HttpClients.custom().setDefaultRequestConfig(requestConfig).setConnectionManager(cm).build();
 
 		this.customHeaders = new HashMap<>();
 		this.putHeader("Content-Type", "application/soap+xml;charset=UTF-8");
@@ -89,7 +82,11 @@ public class ResHttpClient {
 
 			HttpResponse response = this.httpClient.execute(this.host, httpPost);
 			int responseCode = response.getStatusLine().getStatusCode();
-			if (responseCode != HttpStatus.SC_OK) {
+			if (responseCode == HttpStatus.SC_INTERNAL_SERVER_ERROR) {
+				String string = IOUtils.toString(is = response.getEntity().getContent());
+				log.error(string);
+				throw new ResHttpRequestResponseException("HTTP Response code: " + responseCode);
+			} else if (responseCode != HttpStatus.SC_OK) {
 				throw new ResHttpRequestResponseException("HTTP Response code: " + responseCode);
 			}
