+++ src/test/resources/repository/NewFile.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?><Encontro xmlns:oe="http://schemas.openehr.org/v1" xmlns="http://schemas.oceanehr.com/templates" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://schemas.oceanehr.com/templates CN1.xsd" template_id="Resumo de consulta ab_CN1_v1"><name><value>Encontro</value></name><language><terminology_id><value>ISO_639-1</value></terminology_id><code_string>pt</code_string></language><territory><terminology_id><value>ISO_3166-1</value></terminology_id><code_string>BR</code_string></territory><category><value>event</value><defining_code><terminology_id><value>openehr</value></terminology_id><code_string>433</code_string></defining_code></category><composer xsi:type="oe:PARTY_SELF"></composer><context><start_time><oe:value>2016-07-09T00:00:00.000-03:00</oe:value></start_time><setting><oe:value>other care</oe:value><oe:defining_code><oe:terminology_id><oe:value>openehr</oe:value></oe:terminology_id><oe:code_string>238</oe:code_string></oe:defining_code></setting></context><Caracterização_da_consulta><name><value>Caracterização da consulta</value></name><Admissão_do_paciente><name><value>Admissão do paciente</value></name><language><terminology_id><value>ISO_639-1</value></terminology_id><code_string>pt</code_string></language><encoding><terminology_id><value>IANA_character-sets</value></terminology_id><code_string>UTF-8</code_string></encoding><subject></subject><data><Tipo_de_atendimento><name><value>Tipo de atendimento</value></name><value><defining_code><terminology_id><value>local</value></terminology_id><code_string>at0.143</code_string></defining_code></value></Tipo_de_atendimento><Identificação_do_profissional><name><value>Identificação do profissional</value></name><CNS><name><value>CNS</value></name><value><oe:value>123456</oe:value></value></CNS><CBO><name><value>CBO</value></name><value><oe:value>654321</oe:value></value></CBO><É_o_responsável_pelo_atendimento_quest_><name><value>É o responsável pelo atendimento</value></name>
+<value><oe:value>true</oe:value></value></É_o_responsável_pelo_atendimento_quest_></Identificação_do_profissional><Data_fslash_hora_da_admissão><name><value>Data/hora da admissão</value></name><value><oe:value>2016-07-09T00:00:00.000-03:00</oe:value></value></Data_fslash_hora_da_admissão><Turno_de_atendimento><name><value>Turno de atendimento</value></name><value><defining_code><terminology_id><value>local</value></terminology_id><code_string>at0.168</code_string></defining_code></value></Turno_de_atendimento></data></Admissão_do_paciente></Caracterização_da_consulta></Encontro>
